requirements.txt
Pillow==10.1.0
Requests==2.31.0
selenium==4.14.0
streamlit==1.24.1
